history: Separate advisory plan validation phases
History plans describe how saved commits and their patch units become
output commits. Advisory validation reports invalid references, missing
ownership, and unsafe dependency crossings.

I need to inspect phase ordering without tracing one large validator.
Dependency evidence, output checks, and diagnostic accumulation currently
obscure which failures prevent later checks.

This commit gives advisory output and inventory checks separate modules,
shares indexed source and dependency evidence, and collects findings in an
ordered diagnostic value. The strict loader imports the relocated
dependency primitives without changing its validation flow.
